db data displays correctly on artisan profiles
