+++ src/app/profiles/page.tsx
@@ -1,212 +1,8 @@
-"use client";
+import { getSellers } from "../lib/data";
+import ProfileInfo from "./ProfileInfo";
 
-import { useState } from "react";
-import Link from "next/link";
-import Image from "next/image";
-import styles from "../profile.module.css";
+export default async function ProfilePage() {
+  const sellers = await getSellers();
 
-export default function ProfileInfo() {
-  const [comments, setComments] = useState<string[]>([]);
-  const [input, setInput] = useState("");
-  const [showSellerForm, setShowSellerForm] = useState(false);
-
-  const artisans = [
-    {
-      name: "Sarah Cox",
-      craft: "Ceramic Pottery",
-      location: "Utah, USA",
-      image: "/images/poterry.webp",
-    },
-    {
-      name: "Maria Santos",
-      craft: "Handwoven Textiles",
-      location: "Cebu, Philippines",
-      image: "/images/handwoven.webp",
-    },
-    {
-      name: "Aisha Aliezar",
-      craft: "Jewelry Making",
-      location: "Marrakesh, Morocco",
-      image: "/images/jewelry-making.jpeg",
-    },
-  ];
-
-  const updates = [
-    {
-      title: "New Collection: Autumn Designs",
-      date: "Sept 15, 2025",
-      description: "Discover warm colors and cozy patterns inspired by fall.",
-    },
-    {
-      title: "Community Workshop",
-      date: "Oct 1, 2025",
-      description:
-        "Join our live crafting session and learn directly from artisans.",
-    },
-    {
-      title: "Collaboration with Local Farmers",
-      date: "Oct 12, 2025",
-      description:
-        "Supporting sustainability with locally sourced raw materials.",
-    },
-  ];
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!input.trim()) return;
-    setComments([input, ...comments]);
-    setInput("");
-  };
-
-  const handleSellerSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    setShowSellerForm(false);
-    alert("Your seller application has been submitted!");
-  };
-
-  return (
-    <section className={styles.profileSection}>
-      {/* Artisan Profiles */}
-      <article className={styles.card}>
-        <h2>👩‍🎨 Artisan Profiles</h2>
-        <p>
-          Meet our talented artisans! Each profile showcases their story, craft,
-          and featured items.
-        </p>
-        <div className={styles.grid}>
-          {artisans.map((artisan, i) => (
-            <div key={i} className={styles.artisanCard}>
-              <div className={styles.imageWrapper}>
-                <Image
-                  src={artisan.image}
-                  alt={artisan.name}
-                  fill
-                  className={styles.image}
-                />
-              </div>
-              <h3>{artisan.name}</h3>
-              <p className={styles.craft}>{artisan.craft}</p>
-              <p className={styles.location}>{artisan.location}</p>
-            </div>
-          ))}
-        </div>
-      </article>
-
-      {/* Crafting Process */}
-      <article className={styles.card}>
-        <h2>🛠️ Crafting Process</h2>
-        <p>
-          Every creation begins with carefully sourced materials emphasizing
-          sustainability and authenticity.
-        </p>
-        <div className={styles.processGrid}>
-          <div>
-            <span className={styles.icon}>🌱</span>
-            <h4>Sourcing</h4>
-            <p>Eco-friendly raw materials</p>
-          </div>
-          <div>
-            <span className={styles.icon}>🔨</span>
-            <h4>Crafting</h4>
-            <p>Handmade with precision</p>
-          </div>
-          <div>
-            <span className={styles.icon}>📦</span>
-            <h4>Finishing</h4>
-            <p>Polish & quality check</p>
-          </div>
-        </div>
-      </article>
-
-      {/* Customer Engagement */}
-      <article className={styles.card}>
-        <h2>💬 Customer Engagement</h2>
-        <p>Share your thoughts or experiences with our artisans!</p>
-        <form onSubmit={handleSubmit} className={styles.commentForm}>
-          <input
-            type="text"
-            value={input}
-            onChange={(e) => setInput(e.target.value)}
-            placeholder="Write a comment..."
-          />
-          <button type="submit">🚀 Post</button>
-        </form>
-        <div className={styles.commentList}>
-          {comments.length === 0 ? (
-            <p className={styles.noComments}>No comments yet. Be the first!</p>
-          ) : (
-            comments.map((c, i) => (
-              <div key={i} className={styles.commentItem}>
-                <span>💡</span>
-                <p>{c}</p>
-              </div>
-            ))
-          )}
-        </div>
-      </article>
-
-      {/* Community Updates */}
-      <article className={styles.card}>
-        <h2>🌍 Community & Updates</h2>
-        <ul className={styles.updateList}>
-          {updates.map((u, i) => (
-            <li key={i}>
-              <h4>{u.title}</h4>
-              <p className={styles.updateDate}>{u.date}</p>
-              <p>{u.description}</p>
-            </li>
-          ))}
-        </ul>
-      </article>
-
-      {/* Buttons */}
-      <div className={styles.actions}>
-        <button onClick={() => setShowSellerForm(true)}>Become a Seller</button>
-        <Link href="/sellerDashboard">
-          <button className={styles.dashboardBtn}>Seller Dashboard</button>
-        </Link>
-      </div>
-
-      {/* Modal */}
-      {showSellerForm && (
-        <div className={styles.modalOverlay}>
-          <div className={styles.modal}>
-            <button
-              onClick={() => setShowSellerForm(false)}
-              className={styles.closeBtn}
-            >
-              &times;
-            </button>
-            <h2>🛍️ Become a Seller</h2>
-            <form onSubmit={handleSellerSubmit}>
-              <label>
-                Shop Name
-                <input type="text" required placeholder="e.g. Maria’s Shop" />
-              </label>
-              <label>
-                Profile Picture
-                <input type="file" accept="image/*" />
-              </label>
-              <label>
-                Short Description
-                <textarea required placeholder="Describe your shop..." />
-              </label>
-              <div className={styles.modalButtons}>
-                <button
-                  type="button"
-                  onClick={() => setShowSellerForm(false)}
-                  className={styles.cancelBtn}
-                >
-                  Cancel
-                </button>
-                <button type="submit" className={styles.submitBtn}>
-                  Submit
-                </button>
-              </div>
-            </form>
-          </div>
-        </div>
-      )}
-    </section>
-  );
+  return <ProfileInfo sellers={sellers} />;
 }
