+++ src/app/seed/route.ts
@@ -13,7 +13,8 @@ export async function GET() {
         name VARCHAR(255) NOT NULL,
         email VARCHAR(255) NOT NULL UNIQUE,
         location VARCHAR(255),
-        profile_image TEXT
+        profile_image TEXT,
+        craft TEXT
       );
     `;
 
@@ -44,16 +45,16 @@ export async function GET() {
 
     // Insert sellers
     await sql`
-      INSERT INTO sellers (name, email, location, profile_image)
+      INSERT INTO sellers (name, email, location, profile_image, craft)
       VALUES 
-        ('Ana Rivera', 'ana@haven.com', 'Granada, Spain', '/images/sellers/ana.jpg'),
-        ('Liam Torres', 'liam@haven.com', 'Guadalajara, Mexico', '/images/sellers/liam.jpg'),
-        ('Sofia Chen', 'sofia@haven.com', 'Tainan, Taiwan', '/images/sellers/sofia.jpg'),
-        ('Noah Patel', 'noah@haven.com', 'Kolkotta, India', '/images/sellers/noah.jpg'),
-        ('Emma Johnson', 'emma@haven.com', 'Hartford, USA', '/images/sellers/emma.jpg'),
-        ('Sarah Cox', 'sarah@haven.com', 'Utah, USA', '/images/sellers/sarah.jpg'),
-        ('Maria Santos', 'maria@haven.com', 'Cebu, Philippines', '/images/sellers/maria.jpg'),
-        ('Aisha Aliezar', 'aisha@haven.com', 'Marrakesh, Morocco', '/images/sellers/aisha.jpg');
+        ('Ana Rivera', 'ana@haven.com', 'Granada, Spain', '/images/sellers/ana.jpg', 'Ceramic Pottery'),
+        ('Liam Torres', 'liam@haven.com', 'Guadalajara, Mexico', '/images/sellers/liam.jpg', 'Leather Craft'),
+        ('Sofia Chen', 'sofia@haven.com', 'Tainan, Taiwan', '/images/sellers/sofia.jpg', 'Textile Weaving'),
+        ('Noah Patel', 'noah@haven.com', 'Kolkotta, India', '/images/sellers/noah.jpg', 'Ceramic Pottery'),
+        ('Emma Johnson', 'emma@haven.com', 'Hartford, USA', '/images/sellers/emma.jpg', 'Jewelry Making'),
+        ('Sarah Cox', 'sarah@haven.com', 'Utah, USA', '/images/sellers/sarah.jpg', 'Ceramic Pottery'),
+        ('Maria Santos', 'maria@haven.com', 'Cebu, Philippines', '/images/sellers/maria.jpg', 'Textile Weaving'),
+        ('Aisha Aliezar', 'aisha@haven.com', 'Marrakesh, Morocco', '/images/sellers/aisha.jpg', 'Jewelry Making');
     `;
 
     // Insert products (at least 10)
